Consolidate loss functions into a single file (#10737)
* Consolidate loss functions into single file

* updating DIRECTORY.md

* Fix typo

---------
